Create a polygon factory.
We create two functions to create polygons.  One creates regular
polygons and one creates N-pointed stars.

We also add some functions to allow algebra on GeographyValuePoints.  We
have point functions to add and subtract points, to scale points and to
rotate a point around another point through a given angle.  This is
useful in creating the factory functions.

We also add some functions to normalize longitude and latitude values
which are out of the range [-180,180] and [-90,90] respectively.